fix : 헤더 게임 클릭 시 게임 생성 변경
